+++ lib/fossa/views/api.py
@@ -1,6 +1,8 @@
 """
 API views in JSON
 """
+import itertools
+
 from flask import Blueprint, current_app, jsonify, request, url_for
 
 from fossa.control.governor import InvalidTaskSpec
@@ -25,9 +27,11 @@ def test_func(*args):
 
 @api_views.route("/task", methods=["POST"])
 def submit_task():
+    at_capacity_msg = "Node at full capacity and can't accept new tasks"
+
     if not current_app.fossa_governor.has_processing_capacity:
-        # 412 Precondition Failed
-        raise JsonException(message="No spare processing capacity", status_code=412)
+        # 503 Service Unavailable
+        raise JsonException(message=at_capacity_msg, status_code=503)
 
     request_doc = request.get_json()
     if "model_class" not in request_doc:
@@ -50,6 +54,9 @@ def submit_task():
         governor_id = current_app.fossa_governor.submit_task(new_task)
     except InvalidTaskSpec as e:
         raise JsonException(message=str(e), status_code=412)
+
+    if governor_id is None:
+        raise JsonException(message=at_capacity_msg, status_code=503)
 
     api_url = url_for(
         "api.task_details",
@@ -86,10 +93,11 @@ def node_info():
     governor = current_app.fossa_governor
     node_info = node_summary(governor)
 
-    for task in node_info["recent_completed_tasks"]:
+    for task in itertools.chain(node_info["recent_completed_tasks"], node_info["running_tasks"]):
         # remove not serialisable
         for k, v in task.items():
             if callable(v):
                 task[k] = None
 
-    return jsonify(node_info)
+    n_info = jsonify(node_info)
+    return n_info